+++ src/main/java/com/f_log/flog/diet/dto/DailyIntakeDto.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class DailyIntakeDto {
-    private int totalCarbohydrate;
-    private int totalProtein;
-    private int totalFat;
-    private int totalSodium;
-    private int totalCholesterol;
-    private int totalSugars;
-    private int totalCalories;
+    private double totalCarbohydrate;
+    private double totalProtein;
+    private double totalFat;
+    private double totalSodium;
+    private double totalCholesterol;
+    private double totalSugars;
+    private double totalCalories;
 }
